+++ src/constitutive/solid/elastic/almansi_hamel/test.rs
@@ -6,5 +6,3 @@ test_solid_elastic_constitutive_model!(
     ALMANSIHAMELPARAMETERS,
     AlmansiHamel::new(ALMANSIHAMELPARAMETERS)
 );
-
-test_solve_uniaxial!(AlmansiHamel::new(ALMANSIHAMELPARAMETERS));
